refactor(structure): collect interfaces under per-project Interfaces folders
Move all I*.cs interface files in the API into backend/api/Interfaces and
re-namespace them to api.Interfaces. Extract IAuthService from
AuthService.cs into blazor/blazor/Interfaces/IAuthService.cs under
blazor.Interfaces. Update consumers (controllers, services, repositories,
Program.cs, tests, Login.razor) with the new using directives.
