+++ backend/api/Controllers/AssignmentController.cs
@@ -5,6 +5,7 @@ using api.DTOs;
 using api.Models;
 using api.Services;
 using Shared;
+using api.Interfaces;
 
 namespace api.Controllers;
