+++ backend/api/Controllers/AuthController.cs
@@ -4,6 +4,7 @@ using Microsoft.AspNetCore.Cors;
 using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Mvc;
 using Shared.Authorization;
+using api.Interfaces;
 
 namespace api.Controllers;
