+++ backend/api/Controllers/StudentsController.cs
@@ -5,6 +5,7 @@ using FluentValidation;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Shared;
+using api.Interfaces;
 
 namespace api.Controllers;
