+++ backend/api/Interfaces/IAssignmentService.cs
@@ -1,7 +1,7 @@
 using api.DTOs;
 using api.Models;
 
-namespace api.Services;
+namespace api.Interfaces;
 
 public interface IAssignmentService
 {
