+++ backend/api/Interfaces/IUserService.cs
@@ -1,6 +1,6 @@
 using Microsoft.AspNetCore.Identity;
 
-namespace api.Services;
+namespace api.Interfaces;
 
 public interface IUserService
 {
