+++ blazor/blazor/Interfaces/IAuthService.cs
@@ -1,0 +1,7 @@
+namespace blazor.Interfaces;
+
+public interface IAuthService
+{
+    void SetToken(string token);
+    void SignOut();
+}
