+++ blazor/blazor/Pages/Login.razor
@@ -3,6 +3,7 @@
 @inject IAuthService AuthService
 @inject NavigationManager NavManager
 @using Shared.Authorization
+@using blazor.Interfaces
 @using blazor.Services
 
 <EditForm Model="@loginModel" OnValidSubmit="HandleLogin">
