+++ blazor/blazor/Program.cs
@@ -1,4 +1,5 @@
 using blazor;
+using blazor.Interfaces;
 using blazor.Services;
 using Microsoft.AspNetCore.Components.Web;
 using Microsoft.AspNetCore.Components.WebAssembly.Hosting;
